+++ examples/ellipse.py
@@ -1,8 +1,10 @@
+import time
+
 from easygraphics import *
 
 init_graph(600, 600)
 set_render_mode(RenderMode.RENDER_AUTO)
 ellipse(100, 100, 50, 50)
-print(get_line_style())
+time.sleep(2)
 pause()
 close_graph()